# swagger.yaml
basePath: /
consumes:
- application/json
...
info:
  description: Documentation for Product API
  title: Product API.
  version: 1.0.0
paths:
...
  /products/{id}:
    delete:
      description: Delete a Product from database
      operationId: deleteProduct
      parameters:
      - description: the id of the produc to be deleted form the database
        format: int64
        in: path
        name: id
        required: true
        type: integer
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/noContent'
      tags:
      - products
produces:
- application/josn
responses:
  noContent:
    description: ""
...
schemes:
- http
swagger: "2.0"